ai_therapy_app/requirements.txt
# Core Framework
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7

# Database
sqlalchemy==2.0.7
psycopg2-binary==2.9.5
pymongo==4.3.3
redis==4.5.4

# ML and AI
torch==2.0.0
transformers==4.27.4
scikit-learn==1.2.2
spacy==3.5.1
nltk==3.8.1

# Speech and Text Processing
SpeechRecognition==3.10.0
pyttsx3==2.90
openai==0.27.2

# Video Processing
aiortc==1.3.2
av==10.0.0

# Security
cryptography==40.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Task Management
celery==5.2.7

# Testing
pytest==7.3.1
httpx==0.23.3

# Monitoring and Logging
sentry-sdk==1.19.1
python-json-logger==2.0.7

# Utils
python-dotenv==1.0.0
tenacity==8.2.2
requests==2.28.2 